v4 view update and new widgets

+++ EthGasTracker/Models/Historical/HistoricalData.swift
@@ -8,19 +8,19 @@
 import Foundation
 
 struct HistoricalData: Decodable, Equatable {
-    let date: Date  // Now a mandatory field
-    let max: Float
-    let avg: Float
-    let min: Float
-    let p5: Float
-    let p25: Float
-    let p50: Float
-    let p75: Float
-    let p95: Float
+    let date: Date
+    let max:Double
+    let avg:Double
+    let min:Double
+    let p5:Double
+    let p25:Double
+    let p50:Double
+    let p75:Double
+    let p95:Double
     let deviation: Int
     let measureName: String
     
-    var price: Float = 0.0
+    var price:Double = 0.0
     
     enum CodingKeys: String, CodingKey {
         case timestamp, max, avg, min, p5, p25, p50, p75, p95, deviation
@@ -39,14 +39,14 @@ struct HistoricalData: Decodable, Equatable {
         }
         date = parsedDate
         
-        max = try Float(container.decode(String.self, forKey: .max)) ?? 0.0
-        avg = try Float(container.decode(String.self, forKey: .avg)) ?? 0.0
-        min = try Float(container.decode(String.self, forKey: .min)) ?? 0.0
-        p5 = try Float(container.decode(String.self, forKey: .p5)) ?? 0.0
-        p25 = try Float(container.decode(String.self, forKey: .p25)) ?? 0.0
-        p50 = try Float(container.decode(String.self, forKey: .p50)) ?? 0.0
-        p75 = try Float(container.decode(String.self, forKey: .p75)) ?? 0.0
-        p95 = try Float(container.decode(String.self, forKey: .p95)) ?? 0.0
+        max = try Double(container.decode(String.self, forKey: .max)) ?? 0.0
+        avg = try Double(container.decode(String.self, forKey: .avg)) ?? 0.0
+        min = try Double(container.decode(String.self, forKey: .min)) ?? 0.0
+        p5 = try Double(container.decode(String.self, forKey: .p5)) ?? 0.0
+        p25 = try Double(container.decode(String.self, forKey: .p25)) ?? 0.0
+        p50 = try Double(container.decode(String.self, forKey: .p50)) ?? 0.0
+        p75 = try Double(container.decode(String.self, forKey: .p75)) ?? 0.0
+        p95 = try Double(container.decode(String.self, forKey: .p95)) ?? 0.0
         deviation = try Int(container.decode(String.self, forKey: .deviation)) ?? 0
         measureName = try container.decode(String.self, forKey: .measureName)
     }
